+++ src/model.rs
@@ -258,6 +258,7 @@ impl SavedItem {
             extended: self.extended,
             tags,
             dedup_key,
+            toread: false,
         }
     }
 }